pyproject.toml
[project]
name = "notemanager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.32.0",
    "pyzipper>=0.3.6",
    "requests>=2.32.5",
    "rich>=14.1.0",
]
